requirements.txt
gradio
pandas
numpy
plotly
sqlalchemy
mysql-connector-python
python-dotenv
openai